+++ bindings/core/src/method_handler/utils.rs
@@ -1,6 +1,7 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+use crypto::keys::bip39::Mnemonic;
 use iota_sdk::{
     client::{hex_public_key_to_bech32_address, hex_to_bech32, verify_mnemonic, Client},
     types::block::{
@@ -11,7 +12,6 @@ use iota_sdk::{
         Block,
     },
 };
-use zeroize::Zeroize;
 
 use crate::{method::UtilsMethod, response::Response, Result};
 
@@ -29,11 +29,10 @@ pub(crate) fn call_utils_method_internal(method: UtilsMethod) -> Result<Response
         }
         UtilsMethod::ParseBech32Address { address } => Response::ParsedBech32Address(AddressDto::from(address.inner())),
         UtilsMethod::IsAddressValid { address } => Response::Bool(Address::is_valid_bech32(&address)),
-        UtilsMethod::GenerateMnemonic => Response::GeneratedMnemonic(Client::generate_mnemonic()?),
-        UtilsMethod::MnemonicToHexSeed { mut mnemonic } => {
-            let response = Response::MnemonicHexSeed(Client::mnemonic_to_hex_seed(&mnemonic)?);
-            mnemonic.zeroize();
-            response
+        UtilsMethod::GenerateMnemonic => Response::GeneratedMnemonic(Client::generate_mnemonic()?.to_string()),
+        UtilsMethod::MnemonicToHexSeed { mnemonic } => {
+            let mnemonic = Mnemonic::from(mnemonic);
+            Response::MnemonicHexSeed(Client::mnemonic_to_hex_seed(mnemonic)?)
         }
         UtilsMethod::BlockId { block } => {
             let block = Block::try_from_dto_unverified(block)?;
@@ -81,15 +80,15 @@ pub(crate) fn call_utils_method_internal(method: UtilsMethod) -> Result<Response
             let out = Output::try_from_dto_unverified(output)?;
             Response::MinimumRequiredStorageDeposit(out.rent_cost(&rent).to_string())
         }
-        UtilsMethod::VerifyMnemonic { mut mnemonic } => {
-            verify_mnemonic(&mnemonic)?;
-            mnemonic.zeroize();
+        UtilsMethod::VerifyMnemonic { mnemonic } => {
+            let mnemonic = Mnemonic::from(mnemonic);
+            verify_mnemonic(mnemonic)?;
             Response::Ok
         }
         UtilsMethod::VerifyEd25519Signature { signature, message } => {
             let signature = Ed25519Signature::try_from(signature)?;
             let message: Vec<u8> = prefix_hex::decode(message)?;
-            Response::Bool(signature.verify(&message)?)
+            Response::Bool(signature.verify(&message))
         }
         UtilsMethod::VerifySecp256k1EcdsaSignature {
             public_key,
@@ -103,7 +102,7 @@ pub(crate) fn call_utils_method_internal(method: UtilsMethod) -> Result<Response
             let signature = prefix_hex::decode(signature)?;
             let signature = secp256k1_ecdsa::Signature::try_from_bytes(&signature).map_err(Error::from)?;
             let message: Vec<u8> = prefix_hex::decode(message)?;
-            Response::Bool(public_key.verify(&signature, &message))
+            Response::Bool(public_key.verify_keccak256(&signature, &message))
         }
     };
     Ok(response)