+++ sdk/src/client/secret/ledger_nano.rs
@@ -9,7 +9,7 @@ use std::{collections::HashMap, ops::Range};
 
 use async_trait::async_trait;
 use crypto::{
-    keys::slip10::{Chain, Segment},
+    keys::{bip44::Bip44, slip10::Segment},
     signatures::secp256k1_ecdsa::{self, EvmAddress},
 };
 use iota_ledger_nano::{
@@ -141,11 +141,11 @@ impl SecretManage for LedgerSecretManager {
         options: impl Into<Option<GenerateAddressOptions>> + Send,
     ) -> Result<Vec<Ed25519Address>, Self::Error> {
         let options = options.into().unwrap_or_default();
-        let bip32_account = account_index | Segment::HARDEN_MASK;
+        let bip32_account = account_index.harden().into();
 
         let bip32 = LedgerBIP32Index {
-            bip32_index: address_indexes.start | Segment::HARDEN_MASK,
-            bip32_change: u32::from(options.internal) | Segment::HARDEN_MASK,
+            bip32_index: address_indexes.start.harden().into(),
+            bip32_change: u32::from(options.internal).harden().into(),
         };
 
         // lock the mutex to prevent multiple simultaneous requests to a ledger
@@ -177,25 +177,18 @@ impl SecretManage for LedgerSecretManager {
     }
 
     /// Ledger only allows signing messages of 32 bytes, anything else is unsupported and will result in an error.
-    async fn sign_ed25519(&self, msg: &[u8], chain: &Chain) -> Result<Ed25519Signature, Self::Error> {
+    async fn sign_ed25519(&self, msg: &[u8], chain: Bip44) -> Result<Ed25519Signature, Self::Error> {
         if msg.len() != 32 {
             return Err(Error::UnsupportedOperation.into());
         }
 
         let msg = msg.to_vec();
 
-        let bip32_chain = chain
-            .segments()
-            .iter()
-            // XXX: "ser32(i)". RTFSC: [crypto::keys::slip10::Segment::from_u32()]
-            .map(|seg| u32::from_be_bytes(seg.bs()))
-            .collect::<Vec<_>>();
-
-        let coin_type = bip32_chain[1] & !Segment::HARDEN_MASK;
-        let account_index = bip32_chain[2] | Segment::HARDEN_MASK;
+        let coin_type = chain.coin_type;
+        let account_index = chain.account.harden().into();
         let bip32_index = LedgerBIP32Index {
-            bip32_change: bip32_chain[3] | Segment::HARDEN_MASK,
-            bip32_index: bip32_chain[4] | Segment::HARDEN_MASK,
+            bip32_change: chain.change.harden().into(),
+            bip32_index: chain.address_index.harden().into(),
         };
 
         // Lock the mutex to prevent multiple simultaneous requests to a ledger.
@@ -226,7 +219,7 @@ impl SecretManage for LedgerSecretManager {
 
         // Unpack and return signature.
         return match Unlock::unpack::<_, true>(&mut unpacker, &())? {
-            Unlock::Signature(SignatureUnlock(Signature::Ed25519(signature))) => Ok(signature),
+            Unlock::Signature(SignatureUnlock(Signature::Ed25519(signature))) => Ok(*signature),
             _ => Err(Error::UnsupportedOperation.into()),
         };
     }
@@ -234,8 +227,8 @@ impl SecretManage for LedgerSecretManager {
     async fn sign_secp256k1_ecdsa(
         &self,
         _msg: &[u8],
-        _chain: &Chain,
-    ) -> Result<(secp256k1_ecdsa::PublicKey, secp256k1_ecdsa::Signature), Self::Error> {
+        _chain: Bip44,
+    ) -> Result<(secp256k1_ecdsa::PublicKey, secp256k1_ecdsa::RecoverableSignature), Self::Error> {
         Err(Error::UnsupportedOperation.into())
     }
 
@@ -251,39 +244,26 @@ impl SecretManage for LedgerSecretManager {
         let input_len = prepared_transaction.inputs_data.len();
 
         for input in &prepared_transaction.inputs_data {
-            let bip32_indices = match &input.chain {
-                Some(chain) => {
-                    chain
-                        .segments()
-                        .iter()
-                        // XXX: "ser32(i)". RTFSC: [crypto::keys::slip10::Segment::from_u32()]
-                        .map(|seg| u32::from_be_bytes(seg.bs()))
-                        .collect::<Vec<_>>()
-                }
-                None => return Err(Error::MissingBip32Chain)?,
-            };
+            let chain = input.chain.ok_or(Error::MissingBip32Chain)?;
 
             // coin_type and account_index should be the same in each output
-            if (coin_type.is_some() && coin_type != Some(bip32_indices[1]))
-                || (account_index.is_some() && account_index != Some(bip32_indices[2]))
+            if (coin_type.is_some() && coin_type != Some(chain.coin_type))
+                || (account_index.is_some() && account_index != Some(chain.account))
             {
                 return Err(Error::Bip32ChainMismatch.into());
             }
 
-            coin_type = Some(bip32_indices[1]);
-            account_index = Some(bip32_indices[2]);
+            coin_type = Some(chain.coin_type);
+            account_index = Some(chain.account);
             input_bip32_indices.push(LedgerBIP32Index {
-                bip32_change: bip32_indices[3] | Segment::HARDEN_MASK,
-                bip32_index: bip32_indices[4] | Segment::HARDEN_MASK,
+                bip32_change: chain.change.harden().into(),
+                bip32_index: chain.address_index.harden().into(),
             });
         }
 
-        if coin_type.is_none() || account_index.is_none() {
-            return Err(Error::NoAvailableInputsProvided)?;
-        }
+        let (coin_type, account_index) = coin_type.zip(account_index).ok_or(Error::NoAvailableInputsProvided)?;
 
-        let coin_type = coin_type.unwrap() & !Segment::HARDEN_MASK;
-        let bip32_account = account_index.unwrap() | Segment::HARDEN_MASK;
+        let bip32_account = account_index.harden().into();
 
         // pack essence and hash into vec
         let essence_bytes = prepared_transaction.essence.pack_to_vec();
@@ -312,22 +292,12 @@ impl SecretManage for LedgerSecretManager {
             let (remainder_address, remainder_bip32): (Option<&Address>, LedgerBIP32Index) =
                 match &prepared_transaction.remainder {
                     Some(a) => {
-                        let remainder_bip32_indices = match &a.chain {
-                            Some(chain) => {
-                                chain
-                                    .segments()
-                                    .iter()
-                                    // XXX: "ser32(i)". RTFSC: [crypto::keys::slip10::Segment::from_u32()]
-                                    .map(|seg| u32::from_be_bytes(seg.bs()))
-                                    .collect::<Vec<_>>()
-                            }
-                            None => return Err(Error::MissingBip32Chain.into()),
-                        };
+                        let chain = a.chain.ok_or(Error::MissingBip32Chain)?;
                         (
                             Some(&a.address),
                             LedgerBIP32Index {
-                                bip32_change: remainder_bip32_indices[3] | Segment::HARDEN_MASK,
-                                bip32_index: remainder_bip32_indices[4] | Segment::HARDEN_MASK,
+                                bip32_change: chain.change.harden().into(),
+                                bip32_index: chain.address_index.harden().into(),
                             },
                         )
                     }