+++ sdk/src/client/utils.rs
@@ -3,15 +3,16 @@
 
 //! Utility functions for IOTA
 
+use core::borrow::Borrow;
 use std::collections::HashMap;
 
 use crypto::{
     hashes::{blake2b::Blake2b256, Digest},
-    keys::{bip39::wordlist, slip10::Seed},
+    keys::bip39::{wordlist, Mnemonic, MnemonicRef, Passphrase, Seed},
     utils,
 };
 use serde::{Deserialize, Serialize};
-use zeroize::{Zeroize, ZeroizeOnDrop, Zeroizing};
+use zeroize::{Zeroize, ZeroizeOnDrop};
 
 use super::{Client, ClientInner};
 use crate::{
@@ -51,7 +52,7 @@ pub fn hex_public_key_to_bech32_address(hex: &str, bech32_hrp: impl ConvertTo<Hr
 }
 
 /// Generates a new mnemonic.
-pub fn generate_mnemonic() -> Result<String> {
+pub fn generate_mnemonic() -> Result<Mnemonic> {
     let mut entropy = [0u8; 32];
     utils::rand::fill(&mut entropy)?;
     let mnemonic = wordlist::encode(&entropy, &crypto::keys::bip39::wordlist::ENGLISH)
@@ -61,30 +62,23 @@ pub fn generate_mnemonic() -> Result<String> {
 }
 
 /// Returns a hex encoded seed for a mnemonic.
-pub fn mnemonic_to_hex_seed(mnemonic: &str) -> Result<String> {
-    // trim because empty spaces could create a different seed https://github.com/iotaledger/crypto.rs/issues/125
-    let mnemonic = mnemonic.trim();
-    // first we check if the mnemonic is valid to give meaningful errors
-    verify_mnemonic(mnemonic)?;
-    let mut mnemonic_seed = [0u8; 64];
-    crypto::keys::bip39::mnemonic_to_seed(mnemonic, "", &mut mnemonic_seed);
-    Ok(prefix_hex::encode(mnemonic_seed))
+pub fn mnemonic_to_hex_seed(mnemonic: impl Borrow<MnemonicRef>) -> Result<String> {
+    Ok(prefix_hex::encode(mnemonic_to_seed(mnemonic)?.as_ref()))
 }
 
 /// Returns a seed for a mnemonic.
-pub fn mnemonic_to_seed(mnemonic: Zeroizing<String>) -> Result<Seed> {
-    // trim because empty spaces could create a different seed https://github.com/iotaledger/crypto.rs/issues/125
-    let mnemonic = mnemonic.as_str().trim();
+pub fn mnemonic_to_seed(mnemonic: impl Borrow<MnemonicRef>) -> Result<Seed> {
     // first we check if the mnemonic is valid to give meaningful errors
-    verify_mnemonic(mnemonic)?;
-    let mut mnemonic_seed = Zeroizing::new([0u8; 64]);
-    crypto::keys::bip39::mnemonic_to_seed(mnemonic, "", &mut mnemonic_seed);
-    Ok(Seed::from_bytes(mnemonic_seed.as_ref()))
+    verify_mnemonic(mnemonic.borrow())?;
+    Ok(crypto::keys::bip39::mnemonic_to_seed(
+        mnemonic.borrow(),
+        &Passphrase::default(),
+    ))
 }
 
 /// Verifies that a &str is a valid mnemonic.
-pub fn verify_mnemonic(mnemonic: &str) -> Result<()> {
-    crypto::keys::bip39::wordlist::verify(mnemonic, &crypto::keys::bip39::wordlist::ENGLISH)
+pub fn verify_mnemonic(mnemonic: impl Borrow<MnemonicRef>) -> Result<()> {
+    crypto::keys::bip39::wordlist::verify(mnemonic.borrow(), &crypto::keys::bip39::wordlist::ENGLISH)
         .map_err(|e| crate::client::Error::InvalidMnemonic(format!("{e:?}")))?;
     Ok(())
 }
@@ -164,17 +158,17 @@ impl Client {
     }
 
     /// Generates a new mnemonic.
-    pub fn generate_mnemonic() -> Result<String> {
+    pub fn generate_mnemonic() -> Result<Mnemonic> {
         generate_mnemonic()
     }
 
     /// Returns a seed for a mnemonic.
-    pub fn mnemonic_to_seed(mnemonic: Zeroizing<String>) -> Result<Seed> {
+    pub fn mnemonic_to_seed(mnemonic: impl Borrow<MnemonicRef>) -> Result<Seed> {
         mnemonic_to_seed(mnemonic)
     }
 
     /// Returns a hex encoded seed for a mnemonic.
-    pub fn mnemonic_to_hex_seed(mnemonic: &str) -> Result<String> {
+    pub fn mnemonic_to_hex_seed(mnemonic: impl Borrow<MnemonicRef>) -> Result<String> {
         mnemonic_to_hex_seed(mnemonic)
     }