+++ sdk/tests/client/mod.rs
@@ -22,7 +22,7 @@ use std::{
     str::FromStr,
 };
 
-use crypto::keys::slip10::Chain;
+use crypto::keys::bip44::Bip44;
 use iota_sdk::{
     client::secret::types::InputSigningData,
     types::block::{
@@ -71,7 +71,7 @@ enum Build<'a> {
         Option<(&'a str, u64)>,
         Option<u32>,
         Option<(&'a str, u32)>,
-        Option<Chain>,
+        Option<Bip44>,
     ),
     Nft(
         u64,
@@ -82,7 +82,7 @@ enum Build<'a> {
         Option<&'a str>,
         Option<(&'a str, u64)>,
         Option<(&'a str, u32)>,
-        Option<Chain>,
+        Option<Bip44>,
     ),
     Alias(
         u64,
@@ -93,7 +93,7 @@ enum Build<'a> {
         Option<Vec<(&'a str, u64)>>,
         Option<&'a str>,
         Option<&'a str>,
-        Option<Chain>,
+        Option<Bip44>,
     ),
     Foundry(u64, AliasId, u32, SimpleTokenScheme, Option<Vec<(&'a str, u64)>>),
 }
@@ -236,7 +236,7 @@ fn build_foundry_output(
     builder.finish_output(TOKEN_SUPPLY).unwrap()
 }
 
-fn build_output_inner(build: Build) -> (Output, Option<Chain>) {
+fn build_output_inner(build: Build) -> (Output, Option<Bip44>) {
     match build {
         Build::Basic(amount, bech32_address, native_tokens, bech32_sender, sdruc, timelock, expiration, chain) => (
             build_basic_output(